+++ Examples/Mac Demo/iVersionMacAppDelegate.m
@@ -16,15 +16,10 @@
 
 + (void)initialize
 {
-    //set the app and bundle ID. normally you wouldn't need to do this
-    //but we need to test with an app that's actually on the store
-    [iVersion sharedInstance].appStoreID = 412363063;
-    [iVersion sharedInstance].applicationBundleID = @"com.charcoaldesign.RainbowBlocks";
-    
     //configure iVersion. These paths are optional - if you don't set
-    //them, iVersion will just get the release notes from iTunes directly
-	[iVersion sharedInstance].remoteVersionsPlistURL = @"http://charcoaldesign.co.uk/iVersion/versions.plist";
-	[iVersion sharedInstance].localVersionsPlistPath = @"versions.plist";
+    //them, iVersion will just get the release notes from iTunes directly (if your app is on the store)
+    [iVersion sharedInstance].remoteVersionsPlistURL = @"http://charcoaldesign.co.uk/iVersion/versions.plist";
+    [iVersion sharedInstance].localVersionsPlistPath = @"versions.plist";
 }
 
 @end