+++ Examples/iPhone Demo/Classes/iVersionAppDelegate.m
@@ -22,13 +22,8 @@
 
 + (void)initialize
 {
-    //set the app and bundle ID. normally you wouldn't need to do this
-    //but we need to test with an app that's actually on the store
-    [iVersion sharedInstance].appStoreID = 355313284;
-    [iVersion sharedInstance].applicationBundleID = @"com.charcoaldesign.rainbowblocks";
-    
     //configure iVersion. These paths are optional - if you don't set
-    //them, iVersion will just get the release notes from iTunes directly
+    //them, iVersion will just get the release notes from iTunes directly (if your app is on the store)
     [iVersion sharedInstance].remoteVersionsPlistURL = @"http://charcoaldesign.co.uk/iVersion/versions.plist";
     [iVersion sharedInstance].localVersionsPlistPath = @"versions.plist";
 }